Refactor LandscapeEngine.js god file into focused mixin modules
The 1364-line LandscapeEngine.js held a single ~48KB class with
shader code, geometry builders, water mesh setup, terrain chunk
streaming, and the public API all in one place. Split it into five
mixin modules under engine/landscape/ that attach methods to
LandscapeEngine.prototype, leaving the core file as a slim class
skeleton (constructor + public API: setBiome, setStyleMode, clip
bounds, clearChunks, update, dispose).

Layout:
  LandscapeEngine.js           426 lines  (was 1364)
  engine/landscape/noise.js    124 lines  math helpers, getHeight
  engine/landscape/shaders.js  273 lines  sand + low-poly cel shaders
  engine/landscape/geometries  226 lines  rock/flora geos, terrainMat
  engine/landscape/water.js    153 lines  reflective water plane
  engine/landscape/chunks.js   297 lines  chunk builders + queues

The HTML now loads the mixin scripts after LandscapeEngine.js and
before the inline app script, and publish.sh copies the engine/
tree into dist/ alongside the existing static assets.
