+++ tiny-world-builder.html
@@ -4848,6 +4848,11 @@
   <script src="vendor/three/GLTFLoader.r128.js"></script>
   <script src="vendor/tiny-crowd-layer.js"></script>
   <script src="LandscapeEngine.js"></script>
+  <script src="engine/landscape/noise.js"></script>
+  <script src="engine/landscape/shaders.js"></script>
+  <script src="engine/landscape/geometries.js"></script>
+  <script src="engine/landscape/water.js"></script>
+  <script src="engine/landscape/chunks.js"></script>
   <script>
   /* =====================================================================
      Tiny World Builder — single-file Three.js scene + tile editor
